Switch to new authjs setup — broken on Next 14.2
https://www.reddit.com/r/nextjs/comments/1dcc8ti/next_auth_error_could_not_parse_module/
