+++ nextjs/package.json
@@ -68,7 +68,7 @@
     "lucide-react": "^0.344.0",
     "mapbox-gl": "^3.2.0",
     "next": "^14.2.3",
-    "next-auth": "^4.24.7",
+    "next-auth": "^5.0.0-beta.19",
     "next-themes": "^0.2.1",
     "nextjs-toploader": "^1.6.12",
     "node": "^20.14.0",
